API reference

API reference

Visualizing tables

TableWidgets.displaytable — Function.

displaytable(t, rows=1:10; edit = false)

Show rows rows of table t as HTML table. Use : to show the whole table. Use edit=true to make the rows editable. Use reset! to restore original table.

source

TableWidgets.toggletable — Function.

toggletable(t, rows=1:10)

Same as displaytable but the table can be shown or hidden with a toggle switch.

source

Filtering data

TableWidgets.categoricalselector — Function.

categoricalselector(v::AbstractArray, f=filter)

Create as many checkboxes as the unique elements of v and use them to select v. By default it returns a filtered version of v: use categoricalselector(v, map) to get the boolean vector of whether each element is selected

source

TableWidgets.rangeselector — Function.

rangeselector(v::AbstractArray, f=filter)

Create a rangepicker as wide as the extrema of v and uses to select v. By default it returns a filtered version of v: use rangeselector(v, map) to get the boolean vector of whether each element is selected

source

TableWidgets.selector — Function.

selector(v::AbstractArray, f=filter)

Create a textbox where the user can type in an anonymous function that is used to select v. _ can be used to denote the funcion argument, e.g. _ > 0. By default it returns a filtered version of v: use selector(v, map) to get the boolean vector of whether each element is selected

source

TableWidgets.addfilter — Function.

addfilter(t; readout = true)

Create selectors (categoricalselector, rangeselector, selector are supported) and delete them for various columns of table t. readout denotes whether the table will be displayed initially. Outputs the filtered table.

source

Processing data

TableWidgets.dataeditor — Function.

dataeditor(t)

Create a textbox to preprocess a table with JuliaDB / JuliaDBMeta: displays the result using toggletable.

source